Move RowMap into its own file
