+++ src/main/java/com/zendesk/maxwell/MaxwellAbstractRowsEvent.java
@@ -197,47 +197,6 @@ public abstract class MaxwellAbstractRowsEvent extends AbstractRowEvent {
 		return sql.toString();
 	}
 
-	class RowMap extends HashMap<String, Object> {
-		private final HashMap<String, Object> data;
-
-		public RowMap() {
-			this.data = new HashMap<String, Object>();
-			this.put("data", this.data);
-		}
-
-		public void setRowType(String type) {
-			this.put("type", type);
-		}
-
-		public void putData(String key, Object value) {
-			this.data.put(key,  value);
-		}
-
-		public void setTable(String name) {
-			this.put("table", name);
-		}
-
-		public void setDatabase(String name) {
-			this.put("database", name);
-		}
-
-		public void setTimestamp(Long l) {
-			this.put("ts", l);
-		}
-
-		public void setXid(Long xid) {
-			this.put("xid", xid);
-		}
-
-		public void setTXCommit() {
-			this.put("commit", true);
-		}
-
-		public Object getData(String string) {
-			return this.data.get(string);
-		}
-	}
-
 	public List<RowMap> jsonMaps() {
 		ArrayList<RowMap> list = new ArrayList<>();
 		Object value;
@@ -306,7 +265,7 @@ public abstract class MaxwellAbstractRowsEvent extends AbstractRowEvent {
 			g.writeBooleanField("commit", true);
 
 		g.writeObjectFieldStart("data");
-		for ( String key: map.data.keySet() ) {
+		for ( String key: map.dataKeySet() ) {
 			Object data = map.getData(key);
 
 			if ( data == null )
